Feat/Fix:client 함수 재작성 및 연결
